Modify: 코드 다듬기
