+++ nextapp/src/data/store.ts
@@ -2,7 +2,7 @@ import { PostType } from 'types/interface';
 import { create } from 'zustand';
 import { persist, createJSONStorage } from 'zustand/middleware'
 
-export const postStore= create<any>(
+export const postStore= create(
   persist(
     (set, get): any => ({
       nowPost: {},
@@ -19,7 +19,7 @@ export const stateStore: any = create(
   persist(
     (set, get): any => ({
       isLogin: false,
-			
+			darkMode: false,
     }),
     {
       name: 'state-storage', // name of the item in the storage (must be unique)
