+++ nextapp/src/styles/globals.css
@@ -19,6 +19,8 @@
 html {
 	width: 100%;
 	height: 100%;
+	/* background-color: #0d1116 !important; */
+	/* color: #798595; */
 	overflow-y: auto;
 	overflow-x: hidden;
 	background-color: #fafafa;
